# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LearnWise AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized

  # ─── Activity & Progress ────────────────────────────────────────────────────
...
  /study-plans:
    get:
      operationId: getStudyPlans
      tags: [study-plans]
      summary: List study plans
      responses:
        "200":
          description: List of study plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyPlan"
    post:
      operationId: createStudyPlan
      tags: [study-plans]
      summary: Create a study plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanInput"
      responses:
        "201":
          description: Created study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
  /study-plans/{id}:
    get:
      operationId: getStudyPlan
      tags: [study-plans]
      summary: Get a study plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
        "404":
          description: Not found
    patch:
      operationId: updateStudyPlan
      tags: [study-plans]
      summary: Update a study plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyPlanUpdate"
      responses:
        "200":
          description: Updated study plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyPlan"
    delete:
      operationId: deleteStudyPlan
      tags: [study-plans]
      summary: Delete a study plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PDFs ───────────────────────────────────────────────────────────────────
...
  /tutor/sessions:
    get:
      operationId: getTutorSessions
      tags: [tutor]
      summary: List AI tutor sessions
      responses:
        "200":
          description: List of tutor sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TutorSession"
    post:
      operationId: createTutorSession
      tags: [tutor]
      summary: Start a new AI tutor session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorSessionInput"
      responses:
        "201":
          description: Created tutor session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorSession"
  /tutor/sessions/{id}:
    get:
      operationId: getTutorSession
      tags: [tutor]
      summary: Get a tutor session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tutor session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorSession"
        "404":
          description: Not found
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, displayName, email, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        displayName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        studyGoal:
          type: ["string", "null"]
        weeklyGoalHours:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        studyGoal:
          type: string
        weeklyGoalHours:
          type: integer

    # ─── Activity ────────────────────────────────────────────────────────────
...
    StudyPlan:
      type: object
      required: [id, userId, title, subject, durationWeeks, hoursPerWeek, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        subject:
          type: string
        durationWeeks:
          type: integer
        hoursPerWeek:
          type: integer
        status:
          type: string
          enum: [active, completed, paused]
        progressPercent:
          type: integer
        createdAt:
          type: string
          format: date-time
    StudyPlanInput:
      type: object
      required: [title, subject, durationWeeks, hoursPerWeek]
      properties:
        title:
          type: string
        subject:
          type: string
        durationWeeks:
          type: integer
          minimum: 1
        hoursPerWeek:
          type: integer
          minimum: 1
    StudyPlanUpdate:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        durationWeeks:
          type: integer
        hoursPerWeek:
          type: integer
        status:
          type: string
          enum: [active, completed, paused]
        progressPercent:
          type: integer

    # ─── PDFs ────────────────────────────────────────────────────────────────
...
    TutorSession:
      type: object
      required: [id, userId, subject, messageCount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        subject:
          type: string
        messageCount:
          type: integer
        status:
          type: string
          enum: [active, ended]
        lastMessageAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    TutorSessionInput:
      type: object
      required: [subject]
      properties:
        subject:
          type: string