# lib/api-zod/src/generated/types/studyPlan.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LearnWise AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudyPlanStatus } from './studyPlanStatus';

export interface StudyPlan {
  id: number;
  userId: number;
  title: string;
  subject: string;
  durationWeeks: number;
  hoursPerWeek: number;
  status: StudyPlanStatus;
  progressPercent?: number;
  createdAt: Date;
}