# lib/api-zod/src/generated/types/studyPlanUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LearnWise AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudyPlanUpdateStatus } from './studyPlanUpdateStatus';

export interface StudyPlanUpdate {
  title?: string;
  subject?: string;
  durationWeeks?: number;
  hoursPerWeek?: number;
  status?: StudyPlanUpdateStatus;
  progressPercent?: number;
}