# lib/api-zod/src/generated/types/tutorSession.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LearnWise AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TutorSessionStatus } from './tutorSessionStatus';

export interface TutorSession {
  id: number;
  userId: number;
  subject: string;
  messageCount: number;
  status: TutorSessionStatus;
  /** @nullable */
  lastMessageAt?: Date | null;
  createdAt: Date;
}