# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * LearnWise AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileUpdate {
  displayName?: string;
  bio?: string;
  studyGoal?: string;
  weeklyGoalHours?: number;
}